refactor: Renamed graph_utils module.
